Bookshop database project
This script contains the body of the database (8 tables).
The database is designed to resemble an online book store e.g. Amazon.
Project was completed in a team of 4.
Tables include - order table with book price and customer details - newsletter table with subscription types and pricing - bank account table with balance etc.
Another file contains script with queries created to use this database.
Created for CFG SQL course project.